# API_DOC.yaml
openapi: 3.0.0
info:
  title: Hand2Hand API
  version: 1.0.0
...
paths:
...
  /api/user/{id}:
    get:
      tags: [User]
      summary: Get user profile
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: User ID
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  
    put:
      tags: [User]
      summary: Update user profile
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: User ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
  
    delete:
      tags: [User]
      summary: Delete user profile
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: User ID
      responses:
        '200':
          description: User profile deleted
...
  /api/products/:
    get:
      tags: [Products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: condition
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
                  
    post:
      tags: [Products]
      summary: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created
  /api/products/{id}/:
    get:
      tags: [Products]
      summary: Get product detail
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Product ID
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  
    put:
      tags: [Products]
      summary: Edit product
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Product ID
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  
    delete:
      tags: [Products]
      summary: Delete product
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Product ID
      responses:
        '200':
          description: Product deleted
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        user_id:
          type: integer
        last_name:
          type: string
        first_name:
          type: string
        email:
          type: string
        user_type:
          type: string
        is_verified:
          type: boolean
        is_delivery_agent:
          type: boolean
        badge:
          type: string
        sell_count:
          type: integer
        buy_count:
          type: integer
        level:
          type: integer
        bio:
          type: string
        joined_date:
          type: string
          format: date
        profile_pic_url:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        role_id:
          type: number
    Address:
      type: object
      properties:
        address_id:
          type: integer
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
    Product:
      type: object
      properties:
        product_id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        condition:
          type: string
        image_urls:
          type: array
          items:
            type: string
        seller_id:
          type: number
        category_id:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        status:
          type: string
        is_wanted:
          type: boolean
      required:
        - name
        - description
        - price
        - condition
        - image_urls
        - seller_id
        - category_id
        - is_wanted
...